feat: Add first version which passes initial tests.
Add the enum DayOfWeek to the new module `day.py` in `src`.

Add the data classes Hotel and HotelReservation to the new module
`hotel.py` in `src`.

Add the method `get_price_for_reservation(reservation: HotelReservation)`
to the class Hotel which calculates and returns the price for a given
reservation.

Add the functions to get the instances of the three hotels described in
the challenge statement, i.e. Lakewood, Bridgewood and Ridgewood.

Add the first implementation of the function
`get_cheapest_hotel(number: str)` in `my_module.py` which passes the
initial tests.